requirements.txt
pandas~=2.2.3
requests~=2.32.3
lxml~=5.3.2
python-telegram-bot~=22.0